requirements.txt
jsonpickle==0.9.2
webargs==4.0.0
flask==1.0.2
flask-sqlalchemy==2.3.2
thriftpy==0.3.9
flask-script==2.0.6
flask-cors==3.0.6
flask-migrate==2.2.1
gunicorn==19.9.0
blinker==1.4
